+++ src/main/java/com/github/iridatelegrambot/bot/IridaBot.java
@@ -1,14 +1,21 @@
 package com.github.iridatelegrambot.bot;
 
+import com.github.iridatelegrambot.command.*;
+import com.github.iridatelegrambot.service.SendMessageServiceImpl;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.telegram.telegrambots.bots.TelegramLongPollingBot;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
+
+import static com.github.iridatelegrambot.command.CommandName.NO;
 
 @Component
 public class IridaBot extends TelegramLongPollingBot {
+
+    private final static String COMMAND_PREFIX = "/";
+
+    private final CommandContainer container;
 
     @Value("${bot.username}")
     private String username;
@@ -16,6 +23,9 @@ public class IridaBot extends TelegramLongPollingBot {
     @Value("${bot.token}")
     private String token;
 
+    public IridaBot() {
+        this.container = new CommandContainer(new SendMessageServiceImpl(this));
+    }
     @Override
     public String getBotUsername() {
         return username;
@@ -32,15 +42,12 @@ public class IridaBot extends TelegramLongPollingBot {
             String message = update.getMessage().getText().trim();
             String chatId = update.getMessage().getChatId().toString();
 
-            SendMessage sendMessage = new SendMessage();
+            if(message.startsWith(COMMAND_PREFIX)){
+                String commandIdentifier = message.split(" ")[0].toLowerCase();
 
-            sendMessage.setChatId(chatId);
-            sendMessage.setText(message);
-
-            try {
-                execute(sendMessage);
-            } catch (TelegramApiException e) {
-                e.printStackTrace();
+                container.findCommand(commandIdentifier).execute(update);
+            } else {
+                container.findCommand(NO.getCommandName()).execute(update);
             }
         }