+++ src/main/java/com/github/iridatelegrambot/command/Command.java
@@ -1,0 +1,8 @@
+package com.github.iridatelegrambot.command;
+
+import org.telegram.telegrambots.meta.api.objects.Update;
+
+public interface Command {
+
+    void execute(Update update);
+}