+++ RELEASE_NOTES.md
@@ -1,5 +1,10 @@
 # Release Notes
 
+## 0.2.0-SNAPSHOT
+
+*   IRIDA-2: Add telegram bot to the project
+*   IRIDA-3: Add first command for telegram bot
+
 ## 0.1.0-SNAPSHOT
 
-*   JRTB-0: added SpringBoot project
+*   IRIDA-0: create Spring boot project